requirements.txt
pettingzoo==1.24.3
gymnasium==0.29.1
numpy==1.26.4
torch==2.2.0
paho-mqtt==1.6.1
streamlit==1.31.0
pydeck==0.8.1
matplotlib==3.8.2
scipy==1.11.4
pandas==2.2.0
